requirements.txt
plotly==4.14.3
streamlit==0.76.0
pandas==1.1.3
numpy==1.19.5
docx2txt==0.8
pdfplumber==0.5.27
Pillow==8.1.2
scikit_learn==0.24.1
